chore: configure the runtime from the entry instead of the generated module
